+++ src/relstorage/adapters/mysql/schema.py
@@ -24,6 +24,61 @@ from ..schema import AbstractSchemaInstaller
 
 logger = __import__('logging').getLogger(__name__)
 
+# Procedure to set the minimum next OID in one trip to the server.
+# This returns a result, so we'd like to use a function (which we'd access
+# with `SELECT set_min_oid(...)`), but the MySQL manual warns that functions take
+# out table locks and limit concurrency, plus have some replication issues,
+# so we don't do that. Instead we have a procedure. The procedure should be called
+# using `cursor.execute('CALL set_min_oid(%s)` instead of cursor.callproc()
+# --- callproc() involves setting server variables in an extra trip to the server.
+# That would enable us to use an OUT param for the result, but getting that would be another
+# round trip too. Don't forget to use nextset() to get the result of the procedure itself
+# after getting the new oid.
+_SET_MIN_OID = """
+CREATE PROCEDURE set_min_oid(min_oid BIGINT)
+BEGIN
+  -- In order to avoid deadlocks, we only do this if
+  -- the number we want to insert is strictly greater than
+  -- what the current sequence value is. If we use a value less
+  -- than that, there's a chance a different session has already allocated
+  -- and inserted that value into the table, meaning its locked.
+  -- We obviously cannot JUST use MAX(zoid) to find this value, we can't see
+  -- what other sessions have done. But if that's already >= to the min_oid,
+  -- then we don't have to do anything.
+  DECLARE next_oid BIGINT;
+
+  SELECT COALESCE(MAX(ZOID), 0)
+  INTO next_oid
+  FROM new_oid;
+
+  IF next_oid < min_oid THEN
+    -- Can't say for sure. Just because we can only see values
+    -- less doesn't mean they're not there in another transaction.
+
+    -- This will never block.
+    INSERT INTO new_oid VALUES ();
+    SELECT LAST_INSERT_ID()
+    INTO next_oid;
+
+    IF min_oid > next_oid THEN
+      -- This is unlikely to block. We just confirmed that the
+      -- sequence value is strictly less than this, so no one else
+      -- should be doing this.
+      INSERT IGNORE INTO new_oid (zoid)
+      VALUES (min_oid);
+
+      SET next_oid = min_oid;
+    END IF;
+  ELSE
+    -- Return a NULL value to signal that this value cannot
+    -- be cached and used because we didn't allocate it.
+    SET next_oid = NULL;
+  END IF;
+
+  SELECT next_oid;
+END;
+"""
+
 @implementer(ISchemaInstaller)
 class MySQLSchemaInstaller(AbstractSchemaInstaller):
 
@@ -46,10 +101,23 @@ class MySQLSchemaInstaller(AbstractSchemaInstaller):
         'pack_state_tid',
     )
 
+    procedures = {
+        'set_min_oid': _SET_MIN_OID,
+    }
+
+    def __init__(self, driver=None, **kwargs):
+        self.driver = driver
+        super(MySQLSchemaInstaller, self).__init__(**kwargs)
+
     def get_database_name(self, cursor):
         cursor.execute("SELECT DATABASE()")
         for (name,) in cursor:
             return self._metadata_to_native_str(name)
+
+    def list_procedures(self, cursor):
+        cursor.execute("SHOW PROCEDURE STATUS WHERE db = database()")
+        native = self._metadata_to_native_str
+        return [native(row['name']) for row in self._rows_as_dicts(cursor)]
 
     def list_tables(self, cursor):
         return list(self.__list_tables_and_engines(cursor))
@@ -117,7 +185,7 @@ class MySQLSchemaInstaller(AbstractSchemaInstaller):
 
     def _reset_oid(self, cursor):
         from .oidallocator import MySQLOIDAllocator
-        MySQLOIDAllocator().reset_oid(cursor)
+        MySQLOIDAllocator(self.driver).reset_oid(cursor)
 
     def __convert_all_tables_to_innodb(self, cursor):
         tables = self.__list_tables_not_innodb(cursor)
@@ -131,7 +199,15 @@ class MySQLSchemaInstaller(AbstractSchemaInstaller):
         from .oidallocator import MySQLOIDAllocator
         self.__convert_all_tables_to_innodb(cursor)
         super(MySQLSchemaInstaller, self)._prepare_with_connection(conn, cursor)
-        MySQLOIDAllocator().garbage_collect_oids(cursor)
+        MySQLOIDAllocator(self.driver).garbage_collect_oids(cursor)
+
+    def create_procedures(self, cursor):
+        # TODO: Handle updates when we change the text.
+        installed = self.list_procedures(cursor)
+        for name, create_stmt in self.procedures.items():
+            __traceback_info__ = name
+            if name not in installed:
+                cursor.execute(create_stmt)
 
     # We can't TRUNCATE tables that have foreign-key relationships
     # with other tables, but we can drop them. This has to be followed up by
@@ -141,7 +217,7 @@ class MySQLSchemaInstaller(AbstractSchemaInstaller):
     def _after_zap_all_tables(self, cursor, slow=False):
         if not slow:
             logger.debug("Creating tables after drop")
-            self.create(cursor)
+            self.create_tables(cursor)
             logger.debug("Done creating tables after drop")
         else:
             super(MySQLSchemaInstaller, self)._after_zap_all_tables(cursor, slow)