+++ CHANGES.rst
@@ -22,6 +22,11 @@
   returned. Now, it frees the event loop after sending the request.
   See :issue:`272`.
 
+- Call ``set_min_oid`` less often if a storage is just updating
+  existing objects, not creating its own.
+
+- Fix an occasional possible deadlock in MySQL's ``set_min_oid``.
+
 3.0a4 (2019-07-10)
 ==================